refactor(group-members): split monolithic service and controller into focused files
Splits group-members.service.ts (765 lines) and group-members.controller.ts
(295 lines) into three focused service/controller pairs:
- group-members: member removal, role updates, listing
- group-join-requests: submit, accept, reject, withdraw flows
- group-invitations: send, accept, decline, revoke flows

GroupInvitationsController is registered first in the module to prevent
the literal GET /v1/groups/invitations route from being shadowed by the
:id parameterized routes in GroupMembersController.
