+++ apps/api/src/modules/group-members/group-members.controller.ts
@@ -8,11 +8,9 @@ import {
   Param,
   ParseUUIDPipe,
   Patch,
-  Post,
 } from '@nestjs/common';
 import {
   ApiBearerAuth,
-  ApiBody,
   ApiConflictResponse,
   ApiForbiddenResponse,
   ApiNotFoundResponse,
@@ -26,8 +24,6 @@ import {
 import { CurrentUser } from '@/common/decorators/current-user.decorator';
 import type { AuthenticatedUser } from '@/types/express';
 import { GroupMembersService } from './group-members.service';
-import { CreateInvitationDto } from './dto/create-invitation.dto';
-import { BulkInvitationResponseDto } from './dto/bulk-invitation-response.dto';
 import { UpdateMemberRoleDto } from './dto/update-member-role.dto';
 import { MemberResponseDto } from './dto/member-response.dto';
 import { MyMembershipResponseDto } from './dto/my-membership-response.dto';
@@ -38,156 +34,6 @@ import { PendingItemResponseDto } from './dto/pending-item-response.dto';
 @Controller('v1/groups/:id')
 export class GroupMembersController {
   constructor(private readonly groupMembersService: GroupMembersService) {}
-
-  // ─── Join request ─────────────────────────────────────────────────────────────
-
-  @Post('join-request')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Submit a join request',
-    description:
-      'Submits a join request to a public group. Only available when group visibility is PUBLIC.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiResponse({ status: 204, description: 'Join request submitted.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiForbiddenResponse({ description: 'Group is not public.' })
-  @ApiConflictResponse({ description: 'Active membership or pending request already exists.' })
-  @ApiNotFoundResponse({ description: 'Group not found.' })
-  async submitJoinRequest(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): Promise<void> {
-    return this.groupMembersService.submitJoinRequest(id, user.id);
-  }
-
-  @Patch('join-requests/:userId/accept')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Accept a join request',
-    description: 'Accepts a pending join request. Admin only.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiParam({ name: 'userId', type: String, description: 'User UUID of the requester' })
-  @ApiResponse({ status: 204, description: 'Join request accepted.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiForbiddenResponse({ description: 'Caller is not a group admin.' })
-  @ApiConflictResponse({ description: 'No pending join request found.' })
-  @ApiNotFoundResponse({ description: 'Group or user not found.' })
-  async acceptJoinRequest(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Param('userId', ParseUUIDPipe) userId: string,
-  ): Promise<void> {
-    return this.groupMembersService.acceptJoinRequest(id, userId, user.id);
-  }
-
-  @Patch('join-requests/:userId/reject')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Reject a join request',
-    description: 'Rejects a pending join request. Admin only.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiParam({ name: 'userId', type: String, description: 'User UUID of the requester' })
-  @ApiResponse({ status: 204, description: 'Join request rejected.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiForbiddenResponse({ description: 'Caller is not a group admin.' })
-  @ApiConflictResponse({ description: 'No pending join request found.' })
-  @ApiNotFoundResponse({ description: 'Group or user not found.' })
-  async rejectJoinRequest(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Param('userId', ParseUUIDPipe) userId: string,
-  ): Promise<void> {
-    return this.groupMembersService.rejectJoinRequest(id, userId, user.id);
-  }
-
-  // ─── Invitations ──────────────────────────────────────────────────────────────
-
-  @Post('invitations')
-  @HttpCode(HttpStatus.OK)
-  @ApiOperation({
-    summary: 'Send invitations',
-    description:
-      'Sends membership invitations to one or more users by @username. Admin only. ' +
-      'Returns a per-user result for each username in the request.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiBody({ type: CreateInvitationDto })
-  @ApiResponse({
-    status: 200,
-    type: BulkInvitationResponseDto,
-    description: 'Per-user invitation results.',
-  })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiForbiddenResponse({ description: 'Caller is not a group admin.' })
-  async sendInvitations(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Body() dto: CreateInvitationDto,
-  ): Promise<BulkInvitationResponseDto> {
-    return this.groupMembersService.sendInvitations(id, dto, user.id);
-  }
-
-  @Patch('invitations/accept')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Accept an invitation',
-    description: 'Accepts a pending invitation for the authenticated user.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiResponse({ status: 204, description: 'Invitation accepted.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiConflictResponse({ description: 'No pending invitation found.' })
-  @ApiNotFoundResponse({ description: 'Group not found.' })
-  async acceptInvitation(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): Promise<void> {
-    return this.groupMembersService.acceptInvitation(id, user.id);
-  }
-
-  @Patch('invitations/decline')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Decline an invitation',
-    description: 'Declines a pending invitation for the authenticated user.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiResponse({ status: 204, description: 'Invitation declined.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiConflictResponse({ description: 'No pending invitation found.' })
-  @ApiNotFoundResponse({ description: 'Group not found.' })
-  async declineInvitation(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-  ): Promise<void> {
-    return this.groupMembersService.declineInvitation(id, user.id);
-  }
-
-  @Delete('invitations/:userId')
-  @HttpCode(HttpStatus.NO_CONTENT)
-  @ApiOperation({
-    summary: 'Revoke an invitation',
-    description: 'Revokes a pending invitation. Admin only.',
-  })
-  @ApiParam({ name: 'id', type: String, description: 'Group UUID' })
-  @ApiParam({ name: 'userId', type: String, description: 'User UUID of the invitee' })
-  @ApiResponse({ status: 204, description: 'Invitation revoked.' })
-  @ApiUnauthorizedResponse({ description: 'Unauthenticated.' })
-  @ApiForbiddenResponse({ description: 'Caller is not a group admin.' })
-  @ApiConflictResponse({ description: 'No pending invitation found.' })
-  @ApiNotFoundResponse({ description: 'Group or user not found.' })
-  async revokeInvitation(
-    @CurrentUser() user: AuthenticatedUser,
-    @Param('id', ParseUUIDPipe) id: string,
-    @Param('userId', ParseUUIDPipe) userId: string,
-  ): Promise<void> {
-    return this.groupMembersService.revokeInvitation(id, userId, user.id);
-  }
-
-  // ─── Members ──────────────────────────────────────────────────────────────────
 
   @Delete('members/:userId')
   @HttpCode(HttpStatus.NO_CONTENT)
