+++ apps/api/src/modules/group-members/group-members.module.ts
@@ -1,13 +1,16 @@
 import { Module } from '@nestjs/common';
 import { NotificationsModule } from '@/modules/notifications/notifications.module';
-import { GroupMembersController } from './group-members.controller';
 import { GroupInvitationsController } from './group-invitations.controller';
+import { GroupMembersController } from './group-members.controller';
+import { GroupJoinRequestsController } from './group-join-requests.controller';
 import { GroupMembersService } from './group-members.service';
+import { GroupInvitationsService } from './group-invitations.service';
+import { GroupJoinRequestsService } from './group-join-requests.service';
 
 @Module({
   imports: [NotificationsModule],
-  controllers: [GroupInvitationsController, GroupMembersController],
-  providers: [GroupMembersService],
+  controllers: [GroupInvitationsController, GroupMembersController, GroupJoinRequestsController],
+  providers: [GroupMembersService, GroupInvitationsService, GroupJoinRequestsService],
   exports: [GroupMembersService],
 })
 export class GroupMembersModule {}
