+++ apps/api/src/modules/group-members/group-members.service.spec.ts
@@ -12,7 +12,6 @@ import { DRIZZLE_CLIENT } from '@/database/drizzle.provider';
 import { AssetResolverService } from '@/modules/assets/asset-resolver.service';
 import { NotificationsService } from '@/modules/notifications/notifications.service';
 import { GroupMembersService } from './group-members.service';
-import type { CreateInvitationDto } from './dto/create-invitation.dto';
 import type { UpdateMemberRoleDto } from './dto/update-member-role.dto';
 
 jest.mock('@google-cloud/storage', () => ({
@@ -36,8 +35,6 @@ const mockPublicGroup = {
   updatedAt: NOW,
 };
 
-const mockPrivateGroup = { ...mockPublicGroup, visibility: GroupVisibility.PRIVATE };
-
 const makeMembership = (
   userId: string,
   status: GroupMemberStatus,
@@ -57,8 +54,6 @@ const ownerMembership = makeMembership(ADMIN_ID, GroupMemberStatus.ACTIVE, Group
 const activeMembership = makeMembership(USER_ID, GroupMemberStatus.ACTIVE, GroupRole.MEMBER);
 const requestMembership = makeMembership(USER_ID, GroupMemberStatus.REQUEST);
 const invitedMembership = makeMembership(USER_ID, GroupMemberStatus.INVITED);
-const rejectedMembership = makeMembership(USER_ID, GroupMemberStatus.REJECTED);
-const removedMembership = makeMembership(USER_ID, GroupMemberStatus.REMOVED);
 
 const mockTargetUser = {
   id: TARGET_ID,
@@ -183,456 +178,6 @@ describe('GroupMembersService', () => {
     }).compile();
 
     service = module.get<GroupMembersService>(GroupMembersService);
-  });
-
-  // ─── submitJoinRequest ───────────────────────────────────────────────────────
-
-  describe('submitJoinRequest', () => {
-    it('inserts new membership when no existing record', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(undefined); // no existing membership
-
-      await service.submitJoinRequest(GROUP_ID, USER_ID);
-
-      expect(mockInsert).toHaveBeenCalled();
-      expect(mockInsertValues).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.REQUEST, role: GroupRole.MEMBER }),
-      );
-    });
-
-    it('throws NotFoundException when group not found', async () => {
-      mockGroupsFindFirst.mockResolvedValue(undefined);
-
-      await expect(service.submitJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(NotFoundException);
-    });
-
-    it('throws ForbiddenException for private groups', async () => {
-      mockGroupsFindFirst.mockResolvedValue(mockPrivateGroup);
-
-      await expect(service.submitJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(
-        ForbiddenException,
-      );
-    });
-
-    it('throws ConflictException when already ACTIVE', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(activeMembership);
-
-      await expect(service.submitJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(ConflictException);
-    });
-
-    it('throws ConflictException when already REQUEST', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(requestMembership);
-
-      await expect(service.submitJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(ConflictException);
-    });
-
-    it('updates existing row when re-requesting after REJECTED', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(rejectedMembership);
-
-      await service.submitJoinRequest(GROUP_ID, USER_ID);
-
-      expect(mockUpdate).toHaveBeenCalled();
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.REQUEST, role: GroupRole.MEMBER }),
-      );
-      expect(mockInsert).not.toHaveBeenCalled();
-    });
-
-    it('updates existing row when re-requesting after REMOVED', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(removedMembership);
-
-      await service.submitJoinRequest(GROUP_ID, USER_ID);
-
-      expect(mockUpdate).toHaveBeenCalled();
-    });
-  });
-
-  // ─── acceptJoinRequest ───────────────────────────────────────────────────────
-
-  describe('acceptJoinRequest', () => {
-    it('transitions REQUEST → ACTIVE and upserts stats inside transaction', async () => {
-      // Call sequence for groupMembers.findFirst:
-      // 1. assertGroupAdmin → admin membership
-      // 2. findMemberOrThrow → REQUEST membership (target)
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(requestMembership);
-
-      await service.acceptJoinRequest(GROUP_ID, USER_ID, ADMIN_ID);
-
-      expect(mockTransaction).toHaveBeenCalledTimes(1);
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.ACTIVE }),
-      );
-      expect(mockInsertOnConflict).toHaveBeenCalledTimes(1);
-      expect(mockNotificationsNotify).toHaveBeenCalledWith(
-        USER_ID,
-        NotificationType.GROUP_JOIN_ACCEPTED,
-        { groupId: GROUP_ID, groupName: 'Mountain Crew' },
-        [NotificationChannel.PUSH],
-      );
-    });
-
-    it('logs error and resolves when notify() rejects', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(requestMembership);
-      mockNotificationsNotify.mockRejectedValueOnce(new Error('DB blip'));
-      const logSpy = jest.spyOn(service['logger'], 'error').mockImplementation(() => undefined);
-
-      await expect(service.acceptJoinRequest(GROUP_ID, USER_ID, ADMIN_ID)).resolves.toBeUndefined();
-      expect(logSpy).toHaveBeenCalledWith(
-        'Failed to send GROUP_JOIN_ACCEPTED notification',
-        expect.any(Error),
-      );
-    });
-
-    it('throws ForbiddenException when caller is not an admin', async () => {
-      // assertGroupAdmin: group found, but no admin membership
-      mockGroupMembersFindFirst.mockResolvedValueOnce(undefined);
-
-      await expect(service.acceptJoinRequest(GROUP_ID, USER_ID, USER_ID)).rejects.toThrow(
-        ForbiddenException,
-      );
-    });
-
-    it('throws NotFoundException when target has no membership record', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership) // admin check
-        .mockResolvedValueOnce(undefined); // target not found
-
-      await expect(service.acceptJoinRequest(GROUP_ID, TARGET_ID, ADMIN_ID)).rejects.toThrow(
-        NotFoundException,
-      );
-    });
-
-    it('throws ConflictException when target status is not REQUEST', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(invitedMembership); // not a REQUEST
-
-      await expect(service.acceptJoinRequest(GROUP_ID, USER_ID, ADMIN_ID)).rejects.toThrow(
-        ConflictException,
-      );
-    });
-  });
-
-  // ─── rejectJoinRequest ───────────────────────────────────────────────────────
-
-  describe('rejectJoinRequest', () => {
-    it('transitions REQUEST → REJECTED', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(requestMembership);
-
-      await service.rejectJoinRequest(GROUP_ID, USER_ID, ADMIN_ID);
-
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.REJECTED }),
-      );
-    });
-
-    it('throws ConflictException when target is not in REQUEST state', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(activeMembership);
-
-      await expect(service.rejectJoinRequest(GROUP_ID, USER_ID, ADMIN_ID)).rejects.toThrow(
-        ConflictException,
-      );
-    });
-  });
-
-  // ─── withdrawJoinRequest ─────────────────────────────────────────────────────
-
-  describe('withdrawJoinRequest', () => {
-    it('deletes the REQUEST row', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(requestMembership);
-
-      await service.withdrawJoinRequest(GROUP_ID, USER_ID);
-
-      expect(mockDelete).toHaveBeenCalled();
-      expect(mockDeleteWhere).toHaveBeenCalled();
-    });
-
-    it('throws ConflictException when no REQUEST exists', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(activeMembership);
-
-      await expect(service.withdrawJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(
-        ConflictException,
-      );
-    });
-
-    it('throws NotFoundException when membership not found', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(undefined);
-
-      await expect(service.withdrawJoinRequest(GROUP_ID, USER_ID)).rejects.toThrow(
-        NotFoundException,
-      );
-    });
-  });
-
-  // ─── sendInvitations ─────────────────────────────────────────────────────────
-
-  describe('sendInvitations', () => {
-    const dto: CreateInvitationDto = { usernames: ['target_user'] };
-    const targetMembership = (status: GroupMemberStatus, role = GroupRole.MEMBER) =>
-      makeMembership(TARGET_ID, status, role);
-
-    it('inserts new INVITED membership and returns INVITED status', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([]); // no existing membership
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(mockInsert).toHaveBeenCalled();
-      expect(mockInsertValues).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.INVITED }),
-      );
-      expect(result).toEqual({ results: [{ username: 'target_user', status: 'INVITED' }] });
-      expect(mockNotificationsNotifyMany).toHaveBeenCalledWith(
-        [TARGET_ID],
-        NotificationType.GROUP_INVITATION,
-        { groupId: GROUP_ID, groupName: 'Mountain Crew' },
-        [NotificationChannel.PUSH],
-      );
-    });
-
-    it('throws ForbiddenException when caller is not an admin', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(undefined); // no admin membership
-
-      await expect(service.sendInvitations(GROUP_ID, dto, USER_ID)).rejects.toThrow(
-        ForbiddenException,
-      );
-    });
-
-    it('returns NOT_FOUND when target user does not exist', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([]); // user not found
-      mockGroupMembersFindMany.mockResolvedValueOnce([]); // no memberships (no users)
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(result).toEqual({ results: [{ username: 'target_user', status: 'NOT_FOUND' }] });
-      expect(mockInsert).not.toHaveBeenCalled();
-      expect(mockNotificationsNotifyMany).not.toHaveBeenCalled();
-    });
-
-    it('returns ALREADY_MEMBER when target is already ACTIVE', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([targetMembership(GroupMemberStatus.ACTIVE)]);
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(result).toEqual({
-        results: [{ username: 'target_user', status: 'ALREADY_MEMBER' }],
-      });
-      expect(mockNotificationsNotifyMany).not.toHaveBeenCalled();
-    });
-
-    it('returns ALREADY_INVITED when target already has INVITED status', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([targetMembership(GroupMemberStatus.INVITED)]);
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(result).toEqual({
-        results: [{ username: 'target_user', status: 'ALREADY_INVITED' }],
-      });
-      expect(mockNotificationsNotifyMany).not.toHaveBeenCalled();
-    });
-
-    it('returns HAS_PENDING_REQUEST when target has a pending join request', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([targetMembership(GroupMemberStatus.REQUEST)]);
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(result).toEqual({
-        results: [{ username: 'target_user', status: 'HAS_PENDING_REQUEST' }],
-      });
-      expect(mockNotificationsNotifyMany).not.toHaveBeenCalled();
-    });
-
-    it('re-invites after REJECTED, resets role to MEMBER, returns INVITED', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([
-        targetMembership(GroupMemberStatus.REJECTED, GroupRole.ADMIN),
-      ]);
-
-      const result = await service.sendInvitations(GROUP_ID, dto, ADMIN_ID);
-
-      expect(mockUpdate).toHaveBeenCalled();
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.INVITED, role: GroupRole.MEMBER }),
-      );
-      expect(mockInsert).not.toHaveBeenCalled();
-      expect(result).toEqual({ results: [{ username: 'target_user', status: 'INVITED' }] });
-      expect(mockNotificationsNotifyMany).toHaveBeenCalledWith(
-        [TARGET_ID],
-        NotificationType.GROUP_INVITATION,
-        { groupId: GROUP_ID, groupName: 'Mountain Crew' },
-        [NotificationChannel.PUSH],
-      );
-    });
-
-    it('handles multiple usernames and returns per-user results', async () => {
-      const multiDto: CreateInvitationDto = { usernames: ['user_a', 'user_b', 'user_c'] };
-      const userA = { ...mockTargetUser, id: 'id-a', username: 'user_a' };
-      const userB = { ...mockTargetUser, id: 'id-b', username: 'user_b' };
-      const userC = { ...mockTargetUser, id: 'id-c', username: 'user_c' };
-
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership); // admin check
-      mockUsersFindMany.mockResolvedValueOnce([userA, userB, userC]);
-      // userB is already an active member
-      mockGroupMembersFindMany.mockResolvedValueOnce([
-        makeMembership('id-b', GroupMemberStatus.ACTIVE),
-      ]);
-
-      const result = await service.sendInvitations(GROUP_ID, multiDto, ADMIN_ID);
-
-      expect(result).toEqual({
-        results: [
-          { username: 'user_a', status: 'INVITED' },
-          { username: 'user_b', status: 'ALREADY_MEMBER' },
-          { username: 'user_c', status: 'INVITED' },
-        ],
-      });
-      expect(mockNotificationsNotifyMany).toHaveBeenCalledWith(
-        ['id-a', 'id-c'],
-        NotificationType.GROUP_INVITATION,
-        { groupId: GROUP_ID, groupName: 'Mountain Crew' },
-        [NotificationChannel.PUSH],
-      );
-    });
-
-    it('logs error and resolves when notifyMany() rejects', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(ownerMembership);
-      mockUsersFindMany.mockResolvedValueOnce([mockTargetUser]);
-      mockGroupMembersFindMany.mockResolvedValueOnce([]);
-      mockNotificationsNotifyMany.mockRejectedValueOnce(new Error('DB blip'));
-      const logSpy = jest.spyOn(service['logger'], 'error').mockImplementation(() => undefined);
-
-      await expect(service.sendInvitations(GROUP_ID, dto, ADMIN_ID)).resolves.toBeDefined();
-      expect(logSpy).toHaveBeenCalledWith(
-        'Failed to send GROUP_INVITATION notifications',
-        expect.any(Error),
-      );
-    });
-  });
-
-  // ─── acceptInvitation ────────────────────────────────────────────────────────
-
-  describe('acceptInvitation', () => {
-    it('transitions INVITED → ACTIVE, upserts stats, and notifies admins', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(invitedMembership);
-      mockGroupMembersFindMany.mockResolvedValue([{ userId: ADMIN_ID }]);
-      mockUsersFindFirst.mockResolvedValue(mockUserRow); // user accepting the invite
-
-      await service.acceptInvitation(GROUP_ID, USER_ID);
-
-      expect(mockTransaction).toHaveBeenCalledTimes(1);
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.ACTIVE }),
-      );
-      expect(mockInsertOnConflict).toHaveBeenCalledTimes(1);
-      expect(mockNotificationsNotifyMany).toHaveBeenCalledWith(
-        [ADMIN_ID],
-        NotificationType.GROUP_INVITATION_ACCEPTED,
-        expect.objectContaining({
-          groupId: GROUP_ID,
-          groupName: mockPublicGroup.name,
-          username: mockUserRow.username,
-        }),
-        [NotificationChannel.PUSH],
-      );
-    });
-
-    it('does not throw when notifyMany fails', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(invitedMembership);
-      mockGroupMembersFindMany.mockResolvedValue([{ userId: ADMIN_ID }]);
-      mockUsersFindFirst.mockResolvedValue(mockUserRow);
-      mockNotificationsNotifyMany.mockRejectedValue(new Error('FCM error'));
-
-      await expect(service.acceptInvitation(GROUP_ID, USER_ID)).resolves.toBeUndefined();
-    });
-
-    it('skips notification when no active admins exist', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(invitedMembership);
-      mockGroupMembersFindMany.mockResolvedValue([]);
-
-      await service.acceptInvitation(GROUP_ID, USER_ID);
-
-      expect(mockNotificationsNotifyMany).not.toHaveBeenCalled();
-    });
-
-    it('throws ConflictException when no INVITED record exists', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(requestMembership);
-
-      await expect(service.acceptInvitation(GROUP_ID, USER_ID)).rejects.toThrow(ConflictException);
-    });
-
-    it('throws NotFoundException when membership not found', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(undefined);
-
-      await expect(service.acceptInvitation(GROUP_ID, USER_ID)).rejects.toThrow(NotFoundException);
-    });
-  });
-
-  // ─── declineInvitation ───────────────────────────────────────────────────────
-
-  describe('declineInvitation', () => {
-    it('transitions INVITED → REJECTED', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(invitedMembership);
-
-      await service.declineInvitation(GROUP_ID, USER_ID);
-
-      expect(mockUpdateSet).toHaveBeenCalledWith(
-        expect.objectContaining({ status: GroupMemberStatus.REJECTED }),
-      );
-    });
-
-    it('throws ConflictException when no INVITED record exists', async () => {
-      mockGroupMembersFindFirst.mockResolvedValue(activeMembership);
-
-      await expect(service.declineInvitation(GROUP_ID, USER_ID)).rejects.toThrow(ConflictException);
-    });
-  });
-
-  // ─── revokeInvitation ────────────────────────────────────────────────────────
-
-  describe('revokeInvitation', () => {
-    it('deletes the INVITED row', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership) // admin check
-        .mockResolvedValueOnce(invitedMembership); // target
-
-      await service.revokeInvitation(GROUP_ID, USER_ID, ADMIN_ID);
-
-      expect(mockDelete).toHaveBeenCalled();
-    });
-
-    it('throws ForbiddenException when caller is not an admin', async () => {
-      mockGroupMembersFindFirst.mockResolvedValueOnce(undefined);
-
-      await expect(service.revokeInvitation(GROUP_ID, USER_ID, USER_ID)).rejects.toThrow(
-        ForbiddenException,
-      );
-    });
-
-    it('throws ConflictException when target is not INVITED', async () => {
-      mockGroupMembersFindFirst
-        .mockResolvedValueOnce(ownerMembership)
-        .mockResolvedValueOnce(activeMembership);
-
-      await expect(service.revokeInvitation(GROUP_ID, USER_ID, ADMIN_ID)).rejects.toThrow(
-        ConflictException,
-      );
-    });
   });
 
   // ─── removeMember ────────────────────────────────────────────────────────────
