+++ apps/api/src/modules/group-members/group-members.service.ts
@@ -23,16 +23,11 @@ import { NotificationsService } from '@/modules/notifications/notifications.serv
 import { groups } from '@/modules/groups/schema/groups.schema';
 import { groupMembers } from '@/modules/groups/schema/group-members.schema';
 import { groupMemberStats } from '@/modules/groups/schema/group-member-stats.schema';
-import type { CreateInvitationDto } from './dto/create-invitation.dto';
 import type { UpdateMemberRoleDto } from './dto/update-member-role.dto';
 import type { MemberResponseDto } from './dto/member-response.dto';
 import type { PendingItemResponseDto } from './dto/pending-item-response.dto';
 import type { MyMembershipResponseDto } from './dto/my-membership-response.dto';
 import type { MyInvitationResponseDto } from './dto/my-invitation-response.dto';
-import type {
-  BulkInvitationResponseDto,
-  InvitationResultDto,
-} from './dto/bulk-invitation-response.dto';
 
 const ADMIN_ROLES = [GroupRole.OWNER, GroupRole.ADMIN] as const;
 
@@ -45,305 +40,6 @@ export class GroupMembersService {
     private readonly assetResolver: AssetResolverService,
     private readonly notifications: NotificationsService,
   ) {}
-
-  // ─── Join request ────────────────────────────────────────────────────────────
-
-  async submitJoinRequest(groupId: string, requestingUserId: string): Promise<void> {
-    const group = await this.assertGroupExists(groupId);
-    if (group.visibility !== 'PUBLIC') {
-      throw new ForbiddenException('Join requests are only allowed for public groups');
-    }
-
-    const existing = await this.db.query.groupMembers.findFirst({
-      where: and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, requestingUserId)),
-    });
-
-    if (existing) {
-      if (
-        existing.status === GroupMemberStatus.ACTIVE ||
-        existing.status === GroupMemberStatus.REQUEST
-      ) {
-        throw new ConflictException('A pending request or active membership already exists');
-      }
-      // Re-request after REJECTED / REMOVED / LEFT — always reset to MEMBER role
-      await this.db
-        .update(groupMembers)
-        .set({
-          status: GroupMemberStatus.REQUEST,
-          role: GroupRole.MEMBER,
-          initiatedBy: requestingUserId,
-          initiatedAt: new Date(),
-          respondedAt: null,
-          decidedBy: null,
-        })
-        .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, requestingUserId)));
-    } else {
-      await this.db.insert(groupMembers).values({
-        groupId,
-        userId: requestingUserId,
-        status: GroupMemberStatus.REQUEST,
-        role: GroupRole.MEMBER,
-        initiatedBy: requestingUserId,
-      });
-    }
-  }
-
-  async acceptJoinRequest(
-    groupId: string,
-    targetUserId: string,
-    adminUserId: string,
-  ): Promise<void> {
-    await this.assertGroupAdmin(groupId, adminUserId);
-    const membership = await this.findMemberOrThrow(groupId, targetUserId);
-
-    if (membership.status !== GroupMemberStatus.REQUEST) {
-      throw new ConflictException('No pending join request found for this user');
-    }
-
-    const now = new Date();
-    await this.db.transaction(async (trx) => {
-      await trx
-        .update(groupMembers)
-        .set({ status: GroupMemberStatus.ACTIVE, respondedAt: now, decidedBy: adminUserId })
-        .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, targetUserId)));
-
-      await trx
-        .insert(groupMemberStats)
-        .values({ groupId, userId: targetUserId, joinedAt: now })
-        .onConflictDoNothing();
-    });
-
-    const group = await this.db.query.groups.findFirst({
-      where: eq(groups.id, groupId),
-      columns: { name: true },
-    });
-    await this.notifications
-      .notify(
-        targetUserId,
-        NotificationType.GROUP_JOIN_ACCEPTED,
-        { groupId, groupName: group?.name ?? '' },
-        [NotificationChannel.PUSH],
-      )
-      .catch((err: unknown) => {
-        this.logger.error('Failed to send GROUP_JOIN_ACCEPTED notification', err);
-      });
-  }
-
-  async rejectJoinRequest(
-    groupId: string,
-    targetUserId: string,
-    adminUserId: string,
-  ): Promise<void> {
-    await this.assertGroupAdmin(groupId, adminUserId);
-    const membership = await this.findMemberOrThrow(groupId, targetUserId);
-
-    if (membership.status !== GroupMemberStatus.REQUEST) {
-      throw new ConflictException('No pending join request found for this user');
-    }
-
-    await this.db
-      .update(groupMembers)
-      .set({ status: GroupMemberStatus.REJECTED, respondedAt: new Date(), decidedBy: adminUserId })
-      .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, targetUserId)));
-  }
-
-  async withdrawJoinRequest(groupId: string, requestingUserId: string): Promise<void> {
-    const membership = await this.findMemberOrThrow(groupId, requestingUserId);
-
-    if (membership.status !== GroupMemberStatus.REQUEST) {
-      throw new ConflictException('No pending join request to withdraw');
-    }
-
-    await this.db
-      .delete(groupMembers)
-      .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, requestingUserId)));
-  }
-
-  // ─── Invitation ──────────────────────────────────────────────────────────────
-
-  async sendInvitations(
-    groupId: string,
-    dto: CreateInvitationDto,
-    adminUserId: string,
-  ): Promise<BulkInvitationResponseDto> {
-    await this.assertGroupAdmin(groupId, adminUserId);
-
-    const group = await this.db.query.groups.findFirst({
-      where: eq(groups.id, groupId),
-      columns: { name: true },
-    });
-
-    const targetUsers = await this.db.query.users.findMany({
-      where: inArray(users.username, dto.usernames),
-    });
-    const userByUsername = new Map(targetUsers.map((u) => [u.username, u]));
-
-    const existingMemberships =
-      targetUsers.length > 0
-        ? await this.db.query.groupMembers.findMany({
-            where: and(
-              eq(groupMembers.groupId, groupId),
-              inArray(
-                groupMembers.userId,
-                targetUsers.map((u) => u.id),
-              ),
-            ),
-          })
-        : [];
-    const membershipByUserId = new Map(existingMemberships.map((m) => [m.userId, m]));
-
-    const results: InvitationResultDto[] = [];
-    const invitedUserIds: string[] = [];
-
-    for (const username of dto.usernames) {
-      const targetUser = userByUsername.get(username);
-
-      if (!targetUser) {
-        results.push({ username, status: 'NOT_FOUND' });
-        continue;
-      }
-
-      const existing = membershipByUserId.get(targetUser.id);
-
-      if (existing) {
-        if (existing.status === GroupMemberStatus.ACTIVE) {
-          results.push({ username, status: 'ALREADY_MEMBER' });
-          continue;
-        }
-        if (existing.status === GroupMemberStatus.INVITED) {
-          results.push({ username, status: 'ALREADY_INVITED' });
-          continue;
-        }
-        if (existing.status === GroupMemberStatus.REQUEST) {
-          results.push({ username, status: 'HAS_PENDING_REQUEST' });
-          continue;
-        }
-        // REJECTED / REMOVED / LEFT — re-invite, reset to MEMBER role
-        await this.db
-          .update(groupMembers)
-          .set({
-            status: GroupMemberStatus.INVITED,
-            role: GroupRole.MEMBER,
-            initiatedBy: adminUserId,
-            initiatedAt: new Date(),
-            respondedAt: null,
-            decidedBy: null,
-          })
-          .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, targetUser.id)));
-      } else {
-        await this.db.insert(groupMembers).values({
-          groupId,
-          userId: targetUser.id,
-          status: GroupMemberStatus.INVITED,
-          role: GroupRole.MEMBER,
-          initiatedBy: adminUserId,
-        });
-      }
-
-      invitedUserIds.push(targetUser.id);
-      results.push({ username, status: 'INVITED' });
-    }
-
-    if (invitedUserIds.length > 0) {
-      await this.notifications
-        .notifyMany(
-          invitedUserIds,
-          NotificationType.GROUP_INVITATION,
-          { groupId, groupName: group?.name ?? '' },
-          [NotificationChannel.PUSH],
-        )
-        .catch((err: unknown) => {
-          this.logger.error('Failed to send GROUP_INVITATION notifications', err);
-        });
-    }
-
-    return { results };
-  }
-
-  async acceptInvitation(groupId: string, requestingUserId: string): Promise<void> {
-    const membership = await this.findMemberOrThrow(groupId, requestingUserId);
-
-    if (membership.status !== GroupMemberStatus.INVITED) {
-      throw new ConflictException('No pending invitation found');
-    }
-
-    const now = new Date();
-    await this.db.transaction(async (trx) => {
-      await trx
-        .update(groupMembers)
-        .set({ status: GroupMemberStatus.ACTIVE, respondedAt: now, decidedBy: requestingUserId })
-        .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, requestingUserId)));
-
-      await trx
-        .insert(groupMemberStats)
-        .values({ groupId, userId: requestingUserId, joinedAt: now })
-        .onConflictDoNothing();
-    });
-
-    const [group, acceptingUser, adminMembers] = await Promise.all([
-      this.db.query.groups.findFirst({ where: eq(groups.id, groupId), columns: { name: true } }),
-      this.db.query.users.findFirst({
-        where: eq(users.id, requestingUserId),
-        columns: { username: true },
-      }),
-      this.db.query.groupMembers.findMany({
-        where: and(
-          eq(groupMembers.groupId, groupId),
-          eq(groupMembers.status, GroupMemberStatus.ACTIVE),
-          inArray(groupMembers.role, [...ADMIN_ROLES]),
-        ),
-        columns: { userId: true },
-      }),
-    ]);
-
-    const adminIds = adminMembers.map((m) => m.userId);
-    if (adminIds.length > 0) {
-      await this.notifications
-        .notifyMany(
-          adminIds,
-          NotificationType.GROUP_INVITATION_ACCEPTED,
-          { groupId, groupName: group?.name ?? '', username: acceptingUser?.username ?? '' },
-          [NotificationChannel.PUSH],
-        )
-        .catch((err: unknown) => {
-          this.logger.error('Failed to send GROUP_INVITATION_ACCEPTED notification', err);
-        });
-    }
-  }
-
-  async declineInvitation(groupId: string, requestingUserId: string): Promise<void> {
-    const membership = await this.findMemberOrThrow(groupId, requestingUserId);
-
-    if (membership.status !== GroupMemberStatus.INVITED) {
-      throw new ConflictException('No pending invitation found');
-    }
-
-    await this.db
-      .update(groupMembers)
-      .set({
-        status: GroupMemberStatus.REJECTED,
-        respondedAt: new Date(),
-        decidedBy: requestingUserId,
-      })
-      .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, requestingUserId)));
-  }
-
-  async revokeInvitation(
-    groupId: string,
-    targetUserId: string,
-    adminUserId: string,
-  ): Promise<void> {
-    await this.assertGroupAdmin(groupId, adminUserId);
-    const membership = await this.findMemberOrThrow(groupId, targetUserId);
-
-    if (membership.status !== GroupMemberStatus.INVITED) {
-      throw new ConflictException('No pending invitation to revoke');
-    }
-
-    await this.db
-      .delete(groupMembers)
-      .where(and(eq(groupMembers.groupId, groupId), eq(groupMembers.userId, targetUserId)));
-  }
 
   // ─── Remove / Leave ───────────────────────────────────────────────────────────
 
@@ -663,9 +359,9 @@ export class GroupMembersService {
     });
   }
 
-  // ─── Private helpers ──────────────────────────────────────────────────────────
+  // ─── Shared helpers (public — used by GroupJoinRequestsService / GroupInvitationsService) ──
 
-  private async findMemberOrThrow(
+  async findMemberOrThrow(
     groupId: string,
     userId: string,
   ): Promise<typeof groupMembers.$inferSelect> {
@@ -676,7 +372,7 @@ export class GroupMembersService {
     return membership;
   }
 
-  private async assertGroupExists(groupId: string): Promise<typeof groups.$inferSelect> {
+  async assertGroupExists(groupId: string): Promise<typeof groups.$inferSelect> {
     const group = await this.db.query.groups.findFirst({
       where: and(eq(groups.id, groupId), isNull(groups.deletedAt)),
     });
@@ -684,7 +380,7 @@ export class GroupMembersService {
     return group;
   }
 
-  private async assertGroupAdmin(groupId: string, userId: string): Promise<void> {
+  async assertGroupAdmin(groupId: string, userId: string): Promise<void> {
     await this.assertGroupExists(groupId);
 
     const membership = await this.db.query.groupMembers.findFirst({
